Remove documentation of the theme option in config.sample.php
After some discussion around this point, generated by
https://github.com/owncloud/documentation/issues/3728, it's been decided
that documentation of the legacy `theme` option is to be deprecated. So,
to do that, it has to be removed from config/config.sample.php, so that
when the config documentation generator is run, that documentation of
the option is also removed from the docs.
This also fixes https://github.com/owncloud/documentation/issues/3728.